+++ src/Auth/CanDirective.php
@@ -13,7 +13,6 @@ use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Support\Arr;
 use Nuwave\Lighthouse\Exceptions\AuthorizationException;
 use Nuwave\Lighthouse\Exceptions\DefinitionException;
-use Nuwave\Lighthouse\Execution\Arguments\ArgumentSet;
 use Nuwave\Lighthouse\Execution\Resolved;
 use Nuwave\Lighthouse\Execution\ResolveInfo;
 use Nuwave\Lighthouse\Schema\AST\DocumentAST;
@@ -143,7 +142,7 @@ GRAPHQL;
                 );
             }
 
-            foreach ($this->modelsToCheck($resolveInfo->argumentSet, $args) as $model) {
+            foreach ($this->modelsToCheck($root, $args, $context, $resolveInfo) as $model) {
                 $this->authorize($gate, $ability, $model, $checkArguments);
             }
 
@@ -160,13 +159,17 @@ GRAPHQL;
      *
      * @return iterable<\Illuminate\Database\Eloquent\Model|class-string<\Illuminate\Database\Eloquent\Model>>
      */
-    protected function modelsToCheck(ArgumentSet $argumentSet, array $args): iterable
+    protected function modelsToCheck($root, array $args, GraphQLContext $context, ResolveInfo $resolveInfo): iterable
     {
         if ($this->directiveArgValue('query')) {
-            return $argumentSet
+            return $resolveInfo
                 ->enhanceBuilder(
                     $this->getModelClass()::query(),
-                    $this->directiveArgValue('scopes', [])
+                    $this->directiveArgValue('scopes', []),
+                    $root,
+                    $args,
+                    $context,
+                    $resolveInfo
                 )
                 ->get();
         }
@@ -179,7 +182,8 @@ GRAPHQL;
 
             $queryBuilder = $this->getModelClass()::query();
 
-            $directivesContainsForceDelete = $argumentSet->directives->contains(
+            $argumentSetDirectives = $resolveInfo->argumentSet->directives;
+            $directivesContainsForceDelete = $argumentSetDirectives->contains(
                 Utils::instanceofMatcher(ForceDeleteDirective::class)
             );
             if ($directivesContainsForceDelete) {
@@ -188,7 +192,7 @@ GRAPHQL;
                 $queryBuilder->withTrashed();
             }
 
-            $directivesContainsRestore = $argumentSet->directives->contains(
+            $directivesContainsRestore = $argumentSetDirectives->contains(
                 Utils::instanceofMatcher(RestoreDirective::class)
             );
             if ($directivesContainsRestore) {
@@ -198,9 +202,13 @@ GRAPHQL;
             }
 
             try {
-                $enhancedBuilder = $argumentSet->enhanceBuilder(
+                $enhancedBuilder = $resolveInfo->enhanceBuilder(
                     $queryBuilder,
                     $this->directiveArgValue('scopes', []),
+                    $root,
+                    $args,
+                    $context,
+                    $resolveInfo,
                     Utils::instanceofMatcher(TrashedDirective::class)
                 );
                 assert($enhancedBuilder instanceof EloquentBuilder);