+++ src/Schema/Directives/AggregateDirective.php
@@ -4,7 +4,6 @@ namespace Nuwave\Lighthouse\Schema\Directives;
 
 use GraphQL\Language\AST\FieldDefinitionNode;
 use GraphQL\Language\AST\ObjectTypeDefinitionNode;
-use GraphQL\Type\Definition\ResolveInfo;
 use Illuminate\Database\Eloquent\Builder as EloquentBuilder;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Query\Builder as QueryBuilder;
@@ -12,6 +11,7 @@ use Nuwave\Lighthouse\Exceptions\DefinitionException;
 use Nuwave\Lighthouse\Execution\BatchLoader\BatchLoaderRegistry;
 use Nuwave\Lighthouse\Execution\BatchLoader\RelationBatchLoader;
 use Nuwave\Lighthouse\Execution\ModelsLoader\AggregateModelsLoader;
+use Nuwave\Lighthouse\Execution\ResolveInfo;
 use Nuwave\Lighthouse\Schema\AST\DocumentAST;
 use Nuwave\Lighthouse\Schema\Values\FieldValue;
 use Nuwave\Lighthouse\Support\Contracts\FieldManipulator;
@@ -96,6 +96,20 @@ GRAPHQL;
 
     public function resolveField(FieldValue $fieldValue): FieldValue
     {
+        $modelArg = $this->directiveArgValue('model');
+        if (is_string($modelArg)) {
+            $fieldValue->setResolver(function ($root, array $args, GraphQLContext $context, ResolveInfo $resolveInfo) use ($modelArg) {
+                $query = $this->namespaceModelClass($modelArg)::query();
+                assert($query instanceof EloquentBuilder);
+
+                $this->makeBuilderDecorator($root, $args, $context, $resolveInfo)($query);
+
+                return $query->{$this->function()}($this->column());
+            });
+
+            return $fieldValue;
+        }
+
         $relation = $this->directiveArgValue('relation');
         if (is_string($relation)) {
             $fieldValue->setResolver(function (Model $parent, array $args, GraphQLContext $context, ResolveInfo $resolveInfo) {
@@ -104,13 +118,13 @@ GRAPHQL;
                         $this->qualifyPath($args, $resolveInfo),
                         [$this->function(), $this->column()]
                     ),
-                    function () use ($resolveInfo): RelationBatchLoader {
+                    function () use ($parent, $args, $context, $resolveInfo): RelationBatchLoader {
                         return new RelationBatchLoader(
                             new AggregateModelsLoader(
                                 $this->relation(),
                                 $this->column(),
                                 $this->function(),
-                                $this->makeBuilderDecorator($resolveInfo)
+                                $this->makeBuilderDecorator($parent, $args, $context, $resolveInfo)
                             )
                         );
                     }
@@ -129,7 +143,7 @@ GRAPHQL;
                 $query = $this->namespaceModelClass($modelArg)::query();
                 assert($query instanceof EloquentBuilder);
 
-                $this->makeBuilderDecorator($resolveInfo)($query);
+                $this->makeBuilderDecorator($root, $args, $context, $resolveInfo)($query);
 
                 return $query->{$this->function()}($this->column());
             });
@@ -148,7 +162,7 @@ GRAPHQL;
                     "The method referenced by the builder argument of the @{$this->name()} directive on {$this->nodeName()} must return a Builder."
                 );
 
-                $this->makeBuilderDecorator($resolveInfo)($query);
+                $this->makeBuilderDecorator($root, $args, $context, $resolveInfo)($query);
 
                 return $query->{$this->function()}($this->column());
             });