+++ src/Schema/Directives/FindDirective.php
@@ -38,7 +38,11 @@ GRAPHQL;
             $results = $resolveInfo
                 ->enhanceBuilder(
                     $this->getModelClass()::query(),
-                    $this->directiveArgValue('scopes', [])
+                    $this->directiveArgValue('scopes', []),
+                    $root,
+                    $args,
+                    $context,
+                    $resolveInfo
                 )
                 ->get();