+++ src/Schema/Directives/RelationDirectiveHelpers.php
@@ -6,6 +6,7 @@ use Illuminate\Database\Eloquent\Builder as EloquentBuilder;
 use Illuminate\Database\Eloquent\Relations\Relation;
 use Illuminate\Database\Query\Builder as QueryBuilder;
 use Nuwave\Lighthouse\Execution\ResolveInfo;
+use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
 
 trait RelationDirectiveHelpers
 {
@@ -27,11 +28,13 @@ trait RelationDirectiveHelpers
     }
 
     /**
+     * @param  array<string, mixed>  $args
+     *
      * @return \Closure(QueryBuilder|\Illuminate\Database\Eloquent\Builder|\Illuminate\Database\Eloquent\Relations\Relation): void
      */
-    protected function makeBuilderDecorator(ResolveInfo $resolveInfo): \Closure
+    protected function makeBuilderDecorator($root, array $args, GraphQLContext $context, ResolveInfo $resolveInfo): \Closure
     {
-        return function (object $builder) use ($resolveInfo): void {
+        return function (object $builder) use ($root, $args, $context, $resolveInfo): void {
             if ($builder instanceof Relation) {
                 $builder = $builder->getQuery();
             }
@@ -39,7 +42,11 @@ trait RelationDirectiveHelpers
 
             $resolveInfo->enhanceBuilder(
                 $builder,
-                $this->scopes()
+                $this->scopes(),
+                $root,
+                $args,
+                $context,
+                $resolveInfo
             );
         };
     }