+++ src/Schema/Directives/WithDirective.php
@@ -11,6 +11,7 @@ use Nuwave\Lighthouse\Execution\ResolveInfo;
 use Nuwave\Lighthouse\Schema\AST\DocumentAST;
 use Nuwave\Lighthouse\Schema\RootType;
 use Nuwave\Lighthouse\Support\Contracts\FieldManipulator;
+use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
 
 class WithDirective extends WithRelationDirective implements FieldManipulator
 {
@@ -45,11 +46,11 @@ GRAPHQL;
     /**
      * @return \Nuwave\Lighthouse\Execution\ModelsLoader\SimpleModelsLoader
      */
-    protected function modelsLoader(ResolveInfo $resolveInfo): ModelsLoader
+    protected function modelsLoader($parent, array $args, GraphQLContext $context, ResolveInfo $resolveInfo): ModelsLoader
     {
         return new SimpleModelsLoader(
             $this->relation(),
-            $this->makeBuilderDecorator($resolveInfo)
+            $this->makeBuilderDecorator($parent, $args, $context, $resolveInfo)
         );
     }
 }