+++ src/Support/Contracts/FieldBuilderDirective.php
@@ -2,14 +2,17 @@
 
 namespace Nuwave\Lighthouse\Support\Contracts;
 
+use GraphQL\Type\Definition\ResolveInfo;
+
 interface FieldBuilderDirective extends Directive
 {
     /**
      * Add additional constraints to the builder.
      *
      * @param  \Illuminate\Database\Query\Builder|\Illuminate\Database\Eloquent\Builder  $builder  the builder used to resolve the field
+     * @param  array<string, mixed>  $args  the arguments that were passed into the field
      *
      * @return \Illuminate\Database\Query\Builder|\Illuminate\Database\Eloquent\Builder the modified builder
      */
-    public function handleFieldBuilder(object $builder): object;
+    public function handleFieldBuilder(object $builder, $root, array $args, GraphQLContext $context, ResolveInfo $resolveInfo): object;
 }