+++ tests/TestCase.php
@@ -175,8 +175,6 @@ GRAPHQL;
         ]);
 
         $config->set('lighthouse.cache.enable', false);
-
-        $config->set('lighthouse.unbox_bensampo_enum_enum_instances', true);
     }
 
     /**